work on parsing gp productions
Add several tests for the parsing of gp productions. Update the parser
to handle most of these.

The next thing to tackle would be allowing conditions with escaped
spaces, but things are already getting crazy, so shelve this for now.
It turns out that in Soar a separate parser is used for GP, located in
cli_gp.cpp. Further research is needed before continuing.